Add migration for a pre-existing index
An index seems to have been manually added to the production database
without a migration. This adds an appropriate migration to add the
index to any environments that do not already have it. It has been
made repeatable so that if it needs to be dropped, it can have a
DROP clause added and automatically be rerun.

Co-authored-by: Chris Clayson <chris.clayson@digital.cabinet-office.gov.uk>