HUB-717: Add migration for inital two days or request ids
A function has been created to copy request_id and session_id data from
the main audit_events table to the new audit_event_session_requests
table.

It's been realised that there is a _lot_ of data that will be copied.
This initial migration is to copy data for a conservative two days of
the initial date period we're interested in. We will observe how this
migration runs and adjust future migrations as appropriate.