HUB-717: Add migration for request_id for April 10-17
This is being done in small increments to prevent any issues with
locking the DB. The previous migration operated on two days worth of
data and the DB instance had no issues. CPU usage rose to 5% and the
time for the migrations to happen was not abnormal. The event system db
dashboard showed no worrying stats.

This migration will operated on a week of data. After this we could step
up to a month.