HUB-717: Extract request_id for May 17 to Sept 17
The previous migration was for one month and the DB handled it like a
champ.

This migration is for four months and will mean that all data since
since the 9th April up until the point this migration is run will be
migrated.

After this I will proceed in 6 month or one year increments from the
first row in the audit_events table up until 9th April.