HUB-717: Add migration for request_id for 20140905 to 20150505
The previous migration was for a four month period and ran without
issues. This increases the time period to eight months.

As we have now migrated the most recent data this will operate on the
earliest time_stamp and later. The default argument to the function,
`beginning` is omitted here so the default is used - the earliest
time_stamp in the table. Having inspected the DB this is 5 September
2014.