HUB-717: Rename migration to fix order
Flyway orders the migrations by the files timestamp in the prefix. The
last migration was created before the previous, due to last minute
changes. This is preventing the migrations from running succesfully with
an error: "RROR: Validate failed: Detected resolved migration not
applied to database: 20200916155000"

This renames the migration to fix the order flyway will try to apply
them.